Search

+++ features/detail/domain/src/main/java/com/detail/domain/model/DetailProduct.kt
@@ -6,7 +6,7 @@ class DetailProduct(
     id: String,
     name: String,
     thumbnail: String,
-    price: Double,
+    price: Int,
     condition: String,
     val images: List<Image>,
     val description: String?,
